# server/nodejs/api/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0.4"
  title: "Syscoin API"
host: "localhost:8001"
...
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /getblock:
    get:
      tags:
      - "General"
      description: "If verbose is false, returns a string that is serialized, hex-encoded\
        \ data for block 'hash'. If verbose is true, returns an Object with information\
        \ about block <hash>."
      operationId: "getblock"
      parameters:
      - name: "hash"
        in: "query"
        required: true
        type: "string"
      - name: "verbose"
        in: "query"
        required: false
        type: "boolean"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/GetBlockResponse"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /getblockchaininfo:
    get:
      tags:
      - "General"
      description: "Returns an object containing various state info regarding block\
        \ chain processing."
      operationId: "getblockchaininfo"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/GetBlockchainInfoResponse"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /getblockcount:
    get:
      tags:
      - "General"
      description: "Returns the number of blocks in the longest block chain."
      operationId: "getblockcount"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            type: "number"
        default:
          description: "Error"
          schema:
            type: "number"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /getinfo:
    get:
      tags:
      - "General"
      description: "Returns an object containing various state info."
      operationId: "getinfo"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Info"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /getmininginfo:
    get:
      tags:
      - "General"
      description: "Returns a json object containing mining-related information."
      operationId: "getmininginfo"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MiningInfo"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /getnetworkinfo:
    get:
      tags:
      - "General"
      description: "Returns a json object containing network-related information."
      operationId: "getnetworkinfo"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/NetworkInfo"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /getpeerinfo:
    get:
      tags:
      - "General"
      description: "Returns data about each connected network node as a json array\
        \ of objects."
      operationId: "getpeerinfo"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/PeerInfoResponse"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /validateaddress:
    post:
      tags:
      - "General"
      description: "Return information about the given syscoin address."
      operationId: "validateaddress"
      parameters:
      - name: "syscoinaddress"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ValidateAddressResponse"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /verifymessage:
    get:
      tags:
      - "General"
      description: "Verify a signed message"
      operationId: "verifymessage"
      parameters:
      - name: "syscoinaddress"
        in: "query"
        description: "The syscoin address to use for the signature."
        required: true
        type: "string"
      - name: "signature"
        in: "query"
        description: "The signature provided by the signer in base 64 encoding (see\
          \ signmessage)."
        required: true
        type: "string"
      - name: "message"
        in: "query"
        description: "The message that was signed."
        required: true
        type: "string"
      responses:
        200:
          description: "If the signature is verified or not."
          schema:
            type: "boolean"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /addmultisigaddress:
    post:
      tags:
      - "General"
      description: "Add a nrequired-to-sign multisignature address to the wallet.\
        \ Each key is a Syscoin address or hex-encoded public key. If 'account' is\
        \ specified (DEPRECATED), assign address to that account."
      operationId: "addmultisigaddress"
      parameters:
      - in: "body"
        name: "request"
        required: true
        schema:
          $ref: "#/definitions/AddMultisigAddressRequest"
      responses:
        200:
          description: "A syscoin address associated with the keys."
          schema:
            type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
...
  /syscoinsendrawtransaction:
    post:
      tags:
      - "General"
      description: "Signed raw transaction (serialized, hex-encoded) sent out to the\
        \ network."
      operationId: "syscoinsendrawtransaction"
      parameters:
      - in: "body"
        name: "hexstring"
        required: true
        schema:
          $ref: "#/definitions/SendRawTransactionRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /syscoindecoderawtransaction:
    post:
      tags:
      - "General"
      description: "Decode raw syscoin transaction (serialized, hex-encoded) and display\
        \ information pertaining to the service that is included in the transactiion\
        \ data output(OP_RETURN)."
      operationId: "syscoindecoderawtransaction"
      parameters:
      - in: "body"
        name: "hexstring"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "object"
            properties: {}
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
...
  /dumpprivkey:
    get:
      tags:
      - "General"
      description: "Reveals the private key corresponding to 'syscoinaddress'. Then\
        \ the importprivkey can be used with this output."
      operationId: "dumpprivkey"
      parameters:
      - name: "address"
        in: "query"
        description: "The syscoin address for the private key"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /dumpwallet:
    get:
      tags:
      - "General"
      description: "Dumps all wallet keys in a human-readable format."
      operationId: "dumpwallet"
      parameters:
      - name: "filename"
        in: "query"
        description: "The filename"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /encryptwallet:
    post:
      tags:
      - "General"
      description: "Encrypts the wallet with 'passphrase'. This is for first time\
        \ encryption. After this, any calls that interact with private keys such as\
        \ sending or signing will require the passphrase to be set prior the making\
        \ these calls. Use the walletpassphrase call for this, and then walletlock\
        \ call. If the wallet is already encrypted, use the walletpassphrasechange\
        \ call. Note that this will shutdown the server."
      operationId: "encryptwallet"
      parameters:
      - in: "body"
        name: "request"
        required: true
        schema:
          $ref: "#/definitions/EncryptWalletRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
...
  /getaccount:
    get:
      tags:
      - "General"
      description: "DEPRECATED. Returns the account associated with the given address."
      operationId: "getaccount"
      parameters:
      - name: "syscoinaddress"
        in: "query"
        description: "The syscoin address for account lookup."
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /getaccountaddress:
    get:
      tags:
      - "General"
      description: "DEPRECATED. Returns the current Syscoin address for receiving\
        \ payments to this account."
      operationId: "getaccountaddress"
      parameters:
      - name: "account"
        in: "query"
        description: "The account name for the address. It can also be set to the\
          \ empty string \"\" to represent the default account. The account does not\
          \ need to exist, it will be created and a new address created  if there\
          \ is no account by the given name."
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /getaddressbalance:
    get:
      tags:
      - "General"
      description: "Returns the balance for addresses or aliases"
      operationId: "getaddressbalance"
      parameters:
      - name: "addresses"
        in: "query"
        required: true
        type: "array"
        items:
          type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "object"
            properties: {}
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /getaddressutxos:
    get:
      tags:
      - "General"
      description: "Returns all unspent outputs for addresses or aliases"
      operationId: "getaddressutxos"
      parameters:
      - name: "addresses"
        in: "query"
        required: true
        type: "array"
        items:
          type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/GetAddressUTXOsEntry"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /getbalance:
    get:
      tags:
      - "General"
      description: "If account is not specified, returns the server's total available\
        \ balance. If account is specified (DEPRECATED), returns the balance in the\
        \ account. Note that the account \"\" is not the same as leaving the parameter\
        \ out. The server total may be different to the balance in the default \"\"\
        \ account."
      operationId: "getbalance"
      parameters:
      - name: "account"
        in: "query"
        description: "DEPRECATED. The selected account, or \"*\" for entire wallet.\
          \ It may be the default account using \"\"."
        required: false
        type: "string"
      - name: "minconf"
        in: "query"
        description: "Only include transactions confirmed at least this many times."
        required: false
        type: "number"
      - name: "addlockconf"
        in: "query"
        description: "Whether to add 5 confirmations to transactions locked via InstantSend"
        required: false
        type: "boolean"
      - name: "includeWatchonly"
        in: "query"
        description: "Also include balance in watchonly addresses (see 'importaddress')"
        required: false
        type: "boolean"
      responses:
        200:
          description: "Success"
          schema:
            type: "number"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /getwalletbalance:
    get:
      tags:
      - "General"
      description: "Returns wallet balance for all accounts. Does not include watch\
        \ only accounts."
      operationId: "getwalletbalance"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            type: "number"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /getnewaddress:
    post:
      tags:
      - "General"
      description: "Returns a new Syscoin address for receiving payments. If 'account'\
        \ is specified (DEPRECATED), it is added to the address book so payments received\
        \ with the address will be credited to 'account'."
      operationId: "getnewaddress"
      parameters:
      - in: "body"
        name: "request"
        required: false
        schema:
          $ref: "#/definitions/GetNewAddressRequest"
      responses:
        200:
          description: "Success"
          schema:
            type: "string"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
  /getreceivedbyaccount:
    get:
      tags:
      - "General"
      description: "DEPRECATED. Returns the total amount received by addresses with\
        \ <account> in transactions with at least [minconf] confirmations."
      operationId: "getreceivedbyaccount"
      parameters:
      - name: "account"
        in: "query"
        description: "The selected account, may be the default account using \"\"."
        required: true
        type: "string"
      - name: "minconf"
        in: "query"
        description: "Only include transactions confirmed at least this many times."
        required: false
        type: "number"
      - name: "addlockconf"
        in: "query"
        description: "Whether to add 5 confirmations to transactions locked via InstantSend"
        required: false
        type: "boolean"
      responses:
        200:
          description: "Success"
          schema:
            type: "number"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
...
  /getunconfirmedbalance:
    get:
      tags:
      - "General"
      description: "Returns the server's total unconfirmed balance"
      operationId: "getunconfirmedbalance"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            type: "number"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
      - token: []
      x-swagger-router-controller: "General"
    x-swagger-router-controller: "rpc"
...
securityDefinitions:
  token:
    type: "apiKey"
    name: "token"
    in: "header"
definitions:
...
  PeerInfoResponse:
    type: "array"
    items:
      $ref: "#/definitions/PeerInfo"
...
  ErrorResponse:
    required:
    - "message"
    properties:
      message:
        type: "string"
...
  SendRawTransactionRequest:
    required:
    - "hexstring"
    properties:
      hexstring:
        type: "string"
      allowhighfees:
        type: "boolean"
      instantsend:
        type: "boolean"
...
  NetworkInfo:
    properties:
      version:
        type: "number"
      subversion:
        type: "string"
      protocolversion:
        type: "number"
      localservices:
        type: "string"
      localrelay:
        type: "boolean"
      timeoffset:
        type: "number"
      connections:
        type: "number"
      networks:
        type: "array"
        items:
          $ref: "#/definitions/NetworkInfoDetails"
      relayfee:
        type: "number"
      localaddresses:
        type: "array"
        items:
          type: "string"
      warnings:
        type: "string"
    example:
      localaddresses:
      - "aeiou"
      protocolversion: 6.02745618307040320615897144307382404804229736328125
      relayfee: 5.63737665663332876420099637471139430999755859375
      subversion: "aeiou"
      timeoffset: 1.46581298050294517310021547018550336360931396484375
      warnings: "aeiou"
      localrelay: true
      networks:
      - proxy: "aeiou"
        limited: true
        proxy_randomize_credentials: true
        name: "aeiou"
        reachable: true
      version: 0.80082819046101150206595775671303272247314453125
      connections: 5.962133916683182377482808078639209270477294921875
      localservices: "aeiou"
...